requirements.txt
anthropic
pygithub
python-dotenv